Changed approach of loading OS-specific role defaults
Previously, defaults for OS-specific role variables were loaded using a mixture
of include_vars and set_fact. The intention behind this approach was to give a
higher priority to host and group vars over the default role values. It was a
hacky workaround because variables set with include_vars and set_fact have a
high variable precedence and do override group and host vars by default.

The new approach uses the lookup filter and an indirecton in the idiomatic
defaults/main.yml to load the OS-specific role defaults. Hence the OS-specific
defaults have exactly the same variable precedence as ordinary default variables
in defaults/main.yml.